common/requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4
ccxt==4.1.77
colorama==0.4.6
tabulate==0.9.0
schedule==1.2.0 